WIP Unit test and DB tests
